move TaskDataSource.kt to separate class
