+++ app/src/main/java/com/qoiu/dailytaskmotivator/CoreModule.kt
@@ -1,8 +1,8 @@
 package com.qoiu.dailytaskmotivator
 
 import android.content.Context
-import com.qoiu.dailytaskmotivator.data.RealmDataSource
 import com.qoiu.dailytaskmotivator.data.SharedData
+import com.qoiu.dailytaskmotivator.data.TaskDataSource
 import com.qoiu.dailytaskmotivator.data.TaskRepository
 import com.qoiu.dailytaskmotivator.domain.MainInteractor
 import com.qoiu.dailytaskmotivator.domain.TaskInteractor
@@ -16,7 +16,7 @@ class CoreModule {
         val resourceProvider = ResourceProvider.Shared(context)
         val stringProvider = ResourceProvider.String(context)
         val realmProvider = RealmProvider.Base(context)
-        val realmSource = RealmDataSource.Base(realmProvider)
+        val realmSource = TaskDataSource(realmProvider)
         val sharedData = SharedData.Gold(resourceProvider.provideSharedPreference())
         val repository = TaskRepository(realmSource, sharedData)
         mainInteractor = MainInteractor.Base(repository)
